# perf: plugin load time