--- a/ape_hardhat/__init__.py
+++ b/ape_hardhat/__init__.py
@@ -4,20 +4,22 @@ implementation written in Node.js).
 """
 
 from ape import plugins
-from ape.api.networks import LOCAL_NETWORK_NAME
-from ape_ethereum.ecosystem import NETWORKS
-
-from .exceptions import HardhatProviderError, HardhatSubprocessError
-from .provider import HardhatForkProvider, HardhatNetworkConfig, HardhatProvider
 
 
 @plugins.register(plugins.Config)
 def config_class():
+    from .provider import HardhatNetworkConfig
+
     return HardhatNetworkConfig
 
 
 @plugins.register(plugins.ProviderPlugin)
 def providers():
+    from ape.api.networks import LOCAL_NETWORK_NAME
+    from ape_ethereum.ecosystem import NETWORKS
+
+    from .provider import HardhatForkProvider, HardhatProvider
+
     yield "ethereum", LOCAL_NETWORK_NAME, HardhatProvider
 
     for network in NETWORKS:
@@ -55,6 +57,17 @@ def providers():
     yield "gnosis", LOCAL_NETWORK_NAME, HardhatProvider
     yield "gnosis", "mainnet-fork", HardhatForkProvider
     yield "gnosis", "chaido-fork", HardhatForkProvider
+
+
+def __getattr__(name: str):
+    if name.endswith("Error"):
+        import ape_hardhat.exceptions as err_module
+
+        return getattr(err_module, name)
+
+    import ape_hardhat.provider as module
+
+    return getattr(module, name)
 
 
 __all__ = [