### What I did

Before:

```
In [1]: %time import ape_hardhat
CPU times: user 1.17 s, sys: 152 ms, total: 1.33 s
Wall time: 1.36 s
```

After:

```
oIn [1]: %time import ape_hardhat
CPU times: user 8.23 ms, sys: 3.43 ms, total: 11.7 ms
Wall time: 11.5 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)